Added parsing abstract methods, but with a twist.
Instead of generating real abstract methods a method with a dummy body is generated. This is because some sub classes have only @inheritDoc annotation and with abstract this would result in errors; method should be generated while these methods with the annotation are ignored. If they would be generated as native final method this could result in other errors because some classes do not specify the @inheritDoc annotation and these methods would be generated resulting in an error that a final method is overriden (native GWT methods must be final as by GWT convention).