Ignore abstract methods and constructors for abstract classes.
Abstract methods can only be used if they have an implementation. A subclass should have a method specified, that overrides the abstract method. For now the abstract method is not generated. This means if a variable is of type of a super (abstract) class you can't use the abstract method, because it doesn't exist on the class. This is not optimal, but for the use case a lot of sub classes don't implement the abstract classes leading to a lot of errors.
For abstract no constructors should be generated as they can't be created directly, only creating a subclass is useful.